requirements/requirements.txt
selenium == 4.1.3
pytest == 6.2.5
pytest-bdd == 5.0.0
pytest-selenium == 2.0.1
pytest-selenium-enhancer == 1.7.1
allure-pytest-bdd == 2.9.45
allure-python-commons == 2.9.45